Function dams_client::api::create_digital_asset_key

pub fn create_digital_asset_key(
    user_id: UserId,
    blockchain: Blockchain,
    permission: impl UsePermission,
    restriction: impl UseRestriction
) -> Result<KeyInfo, DamsClientError>

Generate a new, distributed digital asset key with the given use parameters for the UserId, and compatible with the specified blockchain.

The UserId must be the same user who opened the crate::DamsClient.

Output: If successful, returns the KeyInfo describing the newly created key.