Function dams_client::api::retrieve_audit_log

pub fn retrieve_audit_log(
    user_id: UserId,
    key_id: Option<&KeyId>
) -> Result<String, DamsClientError>

Retrieve the audit log from the key server for a specified asset owner; optionally, filter for logs associated with the specified KeyId.

The audit log includes context about any action requested and/or taken on the digital asset key, including which action was requested and by whom, the date, details about approval or rejection from each key server, the policy engine, and each asset fiduciary (if relevant), and any other relevant details.

The UserId must match the asset owner authenticated in the crate::DamsClient, and if specified, the KeyId must correspond to a key owned by the UserId.

Output: if successful, returns a String representation of the logs.